\name{listdtr-package}
\alias{listdtr-package}
\docType{package}
\title{
\packageTitle{listdtr}
}
\description{
\packageDescription{listdtr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{listdtr}
\packageIndices{listdtr}
Main function is \code{\link{listdtr}}.
Useful functions include \code{\link{build.rule}}, \code{\link{krr}}.
}
\author{
\packageAuthor{listdtr}

Maintainer: \packageMaintainer{listdtr}
}
\references{
  \url{https://arxiv.org/abs/1606.01472}
}
\keyword{ package }
